requirements.txt
notion-client==2.0.0
python-dotenv==1.0.0
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
aiofiles==23.2.1 